add path module to util, move functions from io.spect into it
